Add files via upload
Analyzing the possibilities of a heart attack under various factors using the dataset heart disease analysis from Kaggle. The exploratory data analysis can be combined with the comprehensive analysis (EDA). Using Logistic Regression with Recursive Feature Elimination(RFE) is used for selecting those features (columns) in a training dataset that are more or most relevant in predicting the target variable and the predictions from the model will tell us whether a person is prone to heart attack or not. Implementing Logistic Regression with PCA takes advantage of multicollinearity and combines the highly correlated variables into a set of uncorrelated variables. Finally, I evaluated the performance of the model.